Time: 23 ms (14.04%) | Memory: 82.4 MB (84.10%) - LeetSync

+++ 1605-minimum-number-of-days-to-make-m-bouquets/minimum-number-of-days-to-make-m-bouquets.java
@@ -1,13 +1,13 @@
 class Solution {
     public int minDays(int[] bloomDay, int m, int k) {
-        long minDay=1L,maxDay=0;
+        int minDay=1,maxDay=0;
         for(int bloom :bloomDay)
             maxDay=Math.max(maxDay,bloom);
-        long ans=Long.MAX_VALUE,c,cm;
+        int ans=Integer.MAX_VALUE,c,cm;
 
         while(minDay<=maxDay)
         {
-            long midDay=minDay+(maxDay-minDay)/2;
+            int midDay=minDay+(maxDay-minDay)/2;
             c=k;
             cm=0;
             for(int b : bloomDay)
@@ -28,6 +28,6 @@ class Solution {
             else
                 minDay=midDay+1;
         }
-        return (ans==Long.MAX_VALUE) ? -1 : (int)ans;
+        return (ans==Integer.MAX_VALUE) ? -1 :ans;
     }
 }
